WIP Implement poison status effect

+++ Pokemon Unity/Assets/Scripts/Data/StatusEffects/StatusEffectNonVolatile.cs
@@ -17,7 +17,7 @@ public class StatusEffectNonVolatile : ScriptableObject
     public bool takesEffectOnceWhenLifeTimeEnded = false;
     public bool takesEffectBeforeMoves = false;
     public int damagePerRoundAbsolute = 0;
-    public int damagePerRoundRelativeToMaxHp = 0;
+    public float damagePerRoundRelativeToMaxHp = 0;
     public int damageByMovement = 0;
     public bool preventsMove = true;
     public float chance = 1f;
